[THREADING] Always accept jobs coming from the CommunicationThread! It is crucial this thread continues.
