+++ Source/core/Thread.h
@@ -249,341 +249,6 @@ namespace Core {
         static uint32_t _defaultStackSize;
     };
 
-    class EXTERNAL ThreadPool {
-    public:
-        typedef Core::QueueType< Core::ProxyType<IDispatch> > MessageQueue;
-
-        template<typename IMPLEMENTATION>
-        class JobType {
-        private:
-            enum state : uint8_t {
-                IDLE,
-                SUBMITTED
-            };
-
-            class Worker : public Core::IDispatch {
-            public:
-                Worker() = delete;
-                Worker(const Worker&) = delete;
-                Worker& operator=(const Worker&) = delete;
-
-                Worker(JobType<IMPLEMENTATION>& parent) : _parent(parent) {
-                }
-                ~Worker() override {
-                }
-
-            private:
-                void Dispatch() override {
-                    _parent.Dispatch();
-                }
-
-            private:
-                 JobType<IMPLEMENTATION>& _parent;
-            };
-
-        public:
-            JobType(const JobType<IMPLEMENTATION>& copy) = delete;
-            JobType<IMPLEMENTATION>& operator=(const JobType<IMPLEMENTATION>& RHS) = delete;
-
-            template <typename... Args>
-            JobType(Args&&... args)
-                : _implementation(args...)
-                , _state(IDLE)
-                , _job(*this)
-            {
-                _job.AddRef();
-            }
-            ~JobType()
-            {
-                ASSERT (_state == IDLE);
-                _job.CompositRelease();
-            }
-
-        public:
-            Core::ProxyType<Core::IDispatch> Aquire() {
-
-                state expected = IDLE;
-                Core::ProxyType<Core::IDispatch> result;
-
-                if (_state.compare_exchange_strong(expected, SUBMITTED) == true) {
-                    result = Core::ProxyType<Core::IDispatch>(&_job, &_job);
-                }
-
-                return (result);
-            }
-            Core::ProxyType<Core::IDispatch> Reset() {
-
-                state expected = SUBMITTED;
-                _state.compare_exchange_strong(expected, IDLE);
-
-                return (Core::ProxyType<Core::IDispatch>(&_job, &_job));
-            }
-            operator IMPLEMENTATION& () {
-                return (_implementation);
-            }
-            operator const IMPLEMENTATION& () const {
-                return (_implementation);
-            }
-
-        protected:
-             Core::ProxyType<Core::IDispatch> Forced() {
-
-                state expected = IDLE;
-                _state.compare_exchange_strong(expected, SUBMITTED);
-
-                return (Core::ProxyType<Core::IDispatch>(&_job, &_job));
-            }
-            
-
-        private:
-            void Dispatch()
-            {
-                state expected = SUBMITTED;
-                if (_state.compare_exchange_strong(expected, IDLE) == true) {
-                    _implementation.Dispatch();
-                }
-            }
-
-        private:
-            IMPLEMENTATION _implementation;
-            std::atomic<state> _state;
-            ProxyObject<Worker> _job;
-        };
-
-        class EXTERNAL Minion {
-        public:
-            Minion(const Minion&) = delete;
-            Minion& operator=(const Minion&) = delete;
-
-            Minion(MessageQueue& queue)
-                : _queue(queue)
-                , _adminLock()
-                , _signal(false, false)
-                , _interestCount(0)
-                , _currentRequest()
-                , _runs(0)
-            {
-            }
-            ~Minion()
-            {
-            }
-
-        public:
-            uint32_t Runs() const {
-                return (_runs);
-            }
-            bool IsActive() const {
-                return (_currentRequest.IsValid());
-            }
-            uint32_t Completed (const Core::ProxyType<Core::IDispatch>& job, const uint32_t waitTime) {
-                uint32_t result = Core::ERROR_NONE;
-
-                _adminLock.Lock();
-                Core::InterlockedIncrement(_interestCount);
-                if (_currentRequest != job) {
-                    _adminLock.Unlock();
-                }
-                else {
-                    _adminLock.Unlock();
-                    result = _signal.Lock(waitTime);
-                }
-
-                Core::InterlockedDecrement(_interestCount);
-
-                return(result);
-            }
-            void Process()
-            {
-                while (_queue.Extract(_currentRequest, Core::infinite) == true) {
-
-                    ASSERT(_currentRequest.IsValid() == true);
-
-                    _runs++;
-
-                    _currentRequest->Dispatch();
-                    _currentRequest.Release();
-
-                    // if someone is observing this run, (WaitForCompletion) make sure that
-                    // thread, sees that his object was running and is now completed.
-                    _adminLock.Lock();
-                    if (_interestCount > 0) {
-
-                        _signal.SetEvent();
-
-                        while (_interestCount > 0) {
-                            ::SleepMs(0);
-                        }
-
-                        _signal.ResetEvent();
-                    }
-                    _adminLock.Unlock();
-                }
-            }
-
-        private:
-            MessageQueue& _queue;
-            Core::CriticalSection _adminLock;
-            Core::Event _signal;
-            uint32_t _interestCount;
-            Core::ProxyType<Core::IDispatch> _currentRequest;
-            uint32_t _runs;
-        };
-
-    private:
-        class EXTERNAL Executor : public Core::Thread {
-        public:
-            Executor(const Executor&) = delete;
-            Executor& operator=(const Executor&) = delete;
-
-            Executor(MessageQueue* queue, const uint32_t stackSize, const TCHAR* name)
-                : Core::Thread(stackSize == 0 ? Core::Thread::DefaultStackSize() : stackSize, name)
-                , _minion(*queue)
-            {
-            }
-            ~Executor() override
-            {
-                Thread::Stop();
-                Wait(Core::Thread::STOPPED, Core::infinite);
-            }
-
-        public:
-            uint32_t Runs() const {
-                return (_minion.Runs());
-            }
-            bool IsActive() const {
-                return (_minion.IsActive());
-            }
-            void Run () {
-                Core::Thread::Run();
-            }
-            void Stop () {
-                Core::Thread::Wait(Core::Thread::STOPPED|Core::Thread::BLOCKED, Core::infinite);
-            }
-            Minion& Me() {
-                return (_minion);
-            }
-
-        private:
-            uint32_t Worker() override
-            {
-                _minion.Process();
-                Core::Thread::Block();
-                return (Core::infinite);
-            }
-
-        private:
-            Minion _minion;
-        };
-
-    public:
-        ThreadPool(const ThreadPool& a_Copy) = delete;
-        ThreadPool& operator=(const ThreadPool& a_RHS) = delete;
-
-        ThreadPool(const uint8_t count, const uint32_t stackSize, const uint32_t queueSize) 
-            : _queue(queueSize)
-        {
-            const TCHAR* name = _T("WorkerPool::Thread");
-            for (uint8_t index = 0; index < count; index++) {
-                _units.emplace_back(&_queue, stackSize, name);
-            }
-        }
-        ~ThreadPool() {
-            Stop();
-            _units.clear();
-        }
-
-    public:
-        uint8_t Count() const
-        {
-            return (static_cast<uint8_t>(_units.size()));
-        }
-        uint32_t Pending() const
-        {
-            return (_queue.Length());
-        }
-        void Runs(const uint8_t length, uint32_t* counters) const 
-        {
-            uint8_t count = 0;
-            std::list<Executor>::const_iterator ptr = _units.cbegin();
-            while ((count < length) && (ptr != _units.cend())) { 
-                counters[count] = ptr->Runs();
-                ptr++; 
-                count++; 
-            }
-        }
-        uint8_t Active() const
-        {
-            uint8_t count = 0;
-            std::list<Executor>::const_iterator ptr = _units.cbegin();
-            while (ptr != _units.cend()) 
-            { 
-                if (ptr->IsActive() == true) {
-                    count++;
-                }
-                ptr++; 
-            }
-
-            return (count);
-        }
-        ::ThreadId Id(const uint8_t index) const
-        {
-            uint8_t count = 0;
-            std::list<Executor>::const_iterator ptr = _units.cbegin();
-            while ((index != count) && (ptr != _units.cend())) { ptr++; count++; }
-
-            ASSERT (ptr != _units.cend());
-
-            return (ptr != _units.cend() ? ptr->Id() : 0);
-        }
-        void Submit(const Core::ProxyType<IDispatch>& job, const uint32_t waitTime)
-        {
-            _queue.Insert(job, waitTime);
-        }
-        uint32_t Revoke(const Core::ProxyType<IDispatch>& job, const uint32_t waitTime)
-        {
-            uint32_t result = Core::ERROR_NONE;
-
-            _queue.Remove(job);
-
-            // Check if it is currently being executed and wait till it is done.
-            std::list<Executor>::iterator index = _units.begin();
-
-            while (index != _units.end()) {
-                uint32_t outcome = index->Me().Completed(job, waitTime);
-                if (outcome != Core::ERROR_NONE) {
-                    result = outcome;
-                }
-                index++;
-            }
-
-            return (result);
-        }
-        MessageQueue& Queue() {
-            return (_queue);
-        }
-        void Run()
-        {
-            _queue.Enable();
-            std::list<Executor>::iterator index = _units.begin();
-            while (index != _units.end()) {
-                index->Run();
-                index++;
-            }
-        }
-        void Stop()
-        {
-            _queue.Disable();
-            std::list<Executor>::iterator index = _units.begin();
-            while (index != _units.end()) {
-                index->Stop();
-                index++;
-            }
-        }
-
-   private:
-        MessageQueue _queue;
-        std::list<Executor> _units;
-    };
 }
 } // namespace Core
